# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EventReceipt": {
        "properties": {
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "task_id",
          "workflow_id",
          "duplicate"
        ],
        "title": "EventReceipt",
        "type": "object"
      },
      "EventView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "service_name": {
            "title": "Service Name",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/TaskSource"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "origin",
          "source",
          "external_id",
          "service_name",
          "title",
          "fingerprint",
          "occurred_at",
          "created_at"
        ],
        "title": "EventView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskSource": {
        "enum": [
          "Alert",
          "Ticket",
          "Schedule",
          "State",
          "Prediction",
          "Release",
          "Human",
          "AI"
        ],
        "title": "TaskSource",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WebhookResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventReceipt"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "WebhookResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "in": "cookie",
        "name": "ops_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Weipai AI Ops Brain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/events/{event_id}": {
      "get": {
        "operationId": "event_detail_api_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "请先登录"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "等待版本或操作内容冲突"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "本地存储或 Temporal 暂时不可用"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Event Detail",
        "tags": [
          "任务闭环"
        ]
      }
    },
...
    "/webhooks/{origin}": {
      "post": {
        "operationId": "receive_event_webhooks__origin__post",
        "parameters": [
          {
            "in": "path",
            "name": "origin",
            "required": true,
            "schema": {
              "enum": [
                "prometheus",
                "kubernetes",
                "ops_platform",
                "git",
                "ci",
                "argocd",
                "config_center",
                "cloud",
                "manual",
                "schedule",
                "state",
                "prediction",
                "learning"
              ],
              "title": "Origin",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receive Event",
        "tags": [
          "事件接入"
        ]
      }
    }
  }
}